adding fake clients, refining tests
