+++ pkg/controller/pubsub/pubsub_config_controller.go
@@ -12,6 +12,7 @@ import (
 	"github.com/open-policy-agent/gatekeeper/v3/pkg/logging"
 	"github.com/open-policy-agent/gatekeeper/v3/pkg/mutation"
 	"github.com/open-policy-agent/gatekeeper/v3/pkg/pubsub"
+	prvd "github.com/open-policy-agent/gatekeeper/v3/pkg/pubsub/provider"
 	"github.com/open-policy-agent/gatekeeper/v3/pkg/readiness"
 	"github.com/open-policy-agent/gatekeeper/v3/pkg/util"
 	"github.com/open-policy-agent/gatekeeper/v3/pkg/watch"
@@ -137,9 +138,18 @@ func (r *Reconciler) Reconcile(ctx context.Context, request reconcile.Request) (
 	if err != nil {
 		return reconcile.Result{}, err
 	}
-	err = r.system.UpsertConnection(ctx, config, request.Name, cfg.Data["provider"])
-	if err != nil {
-		return reconcile.Result{}, err
+	if newConnFunc, ok := prvd.List()[cfg.Data["provider"]]; ok {
+		err = r.system.UpsertConnection(ctx, config, request.Name, cfg.Data["provider"], newConnFunc)
+		if err != nil {
+			return reconcile.Result{}, err
+		}
+	} else {
+		err := r.system.CloseConnection(request.Name)
+		if err != nil {
+			return reconcile.Result{Requeue: true}, err
+		}
+		log.Info("removed connection", "name", request.Name)
+		return reconcile.Result{}, fmt.Errorf("pub-sub provider %s is not supported", cfg.Data["provider"])
 	}
 	log.Info("Connection upsert successful", "name", request.Name, "provider", cfg.Data["provider"])
 	return reconcile.Result{}, nil
